==> Analysis/01_CE211_Cleaning.Rmd <==
---
title: "01_CE211_Cleaning"
author: "Pratik Chaudhari"
date: "10/27/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Load Libraries
```{r}
require(tidyverse)
require(lubridate)
require(stringr)
require(readxl)
require(tidyxl)
require(unpivotr)
```

Load Needs and Referrals Data from data folder
```{r}
col_types_needs <- c('numeric','date','text','text','numeric','text','text','text','text','text','text','text','text','text','text','text','text','text','text','text')

col_types_referrals <- c('numeric','date','text','text','numeric','text','text','text','text','text','text','text','text','text','text','text','text')

col_types_monthly_report <- c('numeric','date','text','text','numeric','text','text','text','text','text','text','text','text','text','text','text','text')

needs_20192020 <- read_excel("../Data/Raw/NeedsbyTaxonomy20192020.xlsx",col_types = col_types_needs)
needs_20202021 <- read_excel("../Data/Raw/NeedsbyTaxonomy20202021.xlsx",col_types = col_types_needs)
referrals_20192020 <- read_excel("../Data/Raw/ReferralstoResources20192020.xlsx",col_types = col_types_referrals)
referrals_20202021 <- read_excel("../Data/Raw/ReferralstoResources20202021.xlsx",col_types = col_types_referrals)
```

Deduplicate Rows where row is exact duplicate
```{r}
needs_20192020 <- needs_20192020 %>% distinct()
needs_20202021 <- needs_20202021 %>% distinct()
referrals_20192020 <- referrals_20192020 %>% distinct()
referrals_20202021 <- referrals_20202021 %>% distinct()
```


Merge dataframes
```{r}
needs <- rbind(needs_20192020,needs_20202021)
referrals <- rbind(referrals_20192020,referrals_20202021)
```

Filter Dataframes for Mecklenburg
```{r}
needs <- needs %>% filter(toupper(County) == 'MECKLENBURG')
referrals <- referrals %>% filter(toupper(County) == 'MECKLENBURG')
```

Type convert
```{r}
needs <- needs %>% mutate(Age = as.factor(Age), Gender = as.factor(Gender), Language = as.factor(Language)
                          ,`Military Status` = as.factor(`Military Status`)
                          ,`Disability Status` = as.factor(`Disability Status`)
                          ,`Disability Type` = as.factor(`Disability Type`)
                          ,`Health Insurance` = as.factor(`Health Insurance`)
                          , Need = as.factor(Need));

referrals <- referrals %>% mutate(Age = as.factor(Age), Gender = as.factor(Gender), Language = as.factor(Language)
                          ,`Military Status` = as.factor(`Military Status`)
                          ,`Disability Status` = as.factor(`Disability Status`)
                          ,`Disability Type` = as.factor(`Disability Type`)
                          ,`Health Insurance` = as.factor(`Health Insurance`))
```

Rename columns, add '_'

```{r}
colnames(needs) <- gsub(pattern = ' ',x=colnames(needs),replacement = '_')
colnames(referrals) <- gsub(pattern = ' ',x=colnames(referrals),replacement = '_')
```

Change DV Referral;Literally Homeless to Literally Homeless;DV Referral

```{r}
needs$CE_Screened <- gsub(x=needs$CE_Screened,pattern = 'DV Referral;Literally Homeless',replacement = 'Literally Homeless;DV Referral')
```


Check for ';' in CE_Screened column
```{r}
head(needs %>% filter(str_detect(CE_Screened,";")))
needs <- separate_rows(needs,'CE_Screened',sep=";")
head(needs)
```

Check for ';' in Referral column
```{r}
head(needs %>% filter(str_detect(Referral,";")))
needs <- separate_rows(needs,'Referral',sep=";")
head(needs)
```
Check for ';' in Disability_Type column
```{r}
head(needs %>% filter(str_detect(Disability_Type,";")))
needs <- separate_rows(needs,'Disability_Type',sep=";")
head(needs)
```


Upper All Column Names
```{r}
colnames(needs) <- toupper(colnames(needs))
colnames(referrals) <- toupper(colnames(referrals))

```

Write to Intermediate file
```{r}
write.csv(x = needs,"../Data/Cleaned/01_Needs.csv",row.names = F)
write.csv(x = referrals,"../Data/Cleaned/01_referrals.csv",row.names = F)
```


Monthly Call Report Cleaning
```{r}
skip_rows = 20
skill_english <- "211 Coordinated Entry English Skill"
skill_spanish <- "SKILL: 211 Coordinated Spanish Skill"
col_names <- c("DATE","...2" ,"...3","CALLS" ,"...5","AGENT count" ,"Reporting.Abandon_Disposition count","ABANDONED count" 
               ,"ABANDONED (%rec)","Average HANDLE TIME" ,"Average SPEED OF ANSWER","Average QUEUE WAIT TIME" 
               ,"Average HOLD TIME","SERVICE LEVEL (%rec)" ,"Average CallSurvey.211CSATQ1MoreInfo"
               ,"CallSurvey.211CSATQ1MoreInfo count" ,"Average CallSurvey.211CSATQ2AgentSatisfaction", 
               "CallSurvey.211CSATQ2AgentSatisfaction count" ,"Average CallSurvey.211CSATQ3Recommendation" 
               ,"CallSurvey.211CSATQ3Recommendation count" ,"CallSurvey.CallSurvey.opt_in count","Reporting.Queue Callback count" 
               ,"file_name","sheet_name")
col_names <- str_squish(col_names)
col_types <- c('date','text','text','text','text','text','text','text','text','text','text','text','text','text','text','text','text','text','text','text','text','text')
```


For each sheet perform the below operation

```{r}
file_list <- list.files(path = '../Data/Raw',pattern = '^CE Monthly Report*',full.names = T)

for(file in file_list){
  df <- data.frame(matrix(ncol = 24, nrow = 0))

  sheet_list <- excel_sheets(file)
  sheet_list <- sheet_list[2:length(sheet_list)]
  processed_sheet <- c()
    
  for(sheet in sheet_list){
    CEMonthly_raw <- read_excel(file,sheet = sheet,skip = 20,trim_ws = T,col_types=col_types)
    colnames(CEMonthly_raw) <- str_squish(colnames(CEMonthly_raw))
    cols_missing <- setdiff(col_names,names(CEMonthly_raw))
    CEMonthly_raw[cols_missing] <- NA
    CEMonthly_raw <- CEMonthly_raw %>% fill(...2,.direction = 'down')
    CEMonthly_raw <- CEMonthly_raw [rowSums(is.na(CEMonthly_raw )) != ncol(CEMonthly_raw ),]
    CEMonthly_raw <- CEMonthly_raw %>% mutate(file_name = gsub(x=basename(file_list[1]),pattern=" ",replacement = "_"),sheet_name = sheet)
    CEMonthly_raw <- CEMonthly_raw %>% mutate(...3 = if_else(str_detect(...2,"SKILL:"),...2,as.character(NA)))
    CEMonthly_raw <- CEMonthly_raw %>% fill(...3,.direction = 'down')
    df <- rbind(df,CEMonthly_raw)
    processed_sheet <- c(processed_sheet,sheet)
  }
  print(paste("Writing file",paste('../Data/Cleaned/01_CE_Report_Stage_1/'
                                ,gsub(x = gsub(x=basename(file),pattern=" ",replacement = "_"),pattern = '.xlsx',replacement = '.csv')
                                ,sep = '')))
  write.csv(df,paste('../Data/Cleaned/01_CE_Report_Stage_1/'
                               ,gsub(x = gsub(x=basename(file),pattern=" ",replacement = "_"),pattern = '.xlsx',replacement = '.csv')
                               ,sep = ''),row.names = F)
}
```

Read from Stage 1 Cleaned Monthly Report Files
```{r}
CEMonthly_2019 <- read.csv("../Data/Cleaned/01_CE_Report_Stage_1/CE_Monthly_Report_2019.csv")

CEMonthly_2020 <- read.csv("../Data/Cleaned/01_CE_Report_Stage_1/CE_Monthly_Report_2020.csv")

CEMonthly_2021 <- read.csv("../Data/Cleaned/01_CE_Report_Stage_1/CE_Monthly_Report_2021.csv")

```

Clean Further 
1. Clean Null records and keep only relevant rows 
  Removed aggregated rows
```{r}
CEMonthly_2019 <- CEMonthly_2019[,colSums(is.na(CEMonthly_2019))<nrow(CEMonthly_2019)]
CEMonthly_2020 <- CEMonthly_2020[,colSums(is.na(CEMonthly_2020))<nrow(CEMonthly_2020)]
CEMonthly_2021 <- CEMonthly_2021[,colSums(is.na(CEMonthly_2021))<nrow(CEMonthly_2021)]

CEMonthly_2019 <- CEMonthly_2019 %>% filter(is.na(DATE) == F)
CEMonthly_2020 <- CEMonthly_2020 %>% filter(is.na(DATE) == F)
CEMonthly_2021 <- CEMonthly_2021 %>% filter(is.na(DATE) == F)
```

2. Column data type

```{r}
CEMonthly_2019 <- CEMonthly_2019 %>% mutate_at(.vars = c("CALLS", "AGENT.count","Reporting.Abandon_Disposition.count"
                                                         , "ABANDONED.count", "ABANDONED...rec."
                                                         , "SERVICE.LEVEL...rec.", "Average.CallSurvey.211CSATQ1MoreInfo"
                                                         , "CallSurvey.211CSATQ1MoreInfo.count"
                                                         , "Average.CallSurvey.211CSATQ2AgentSatisfaction"
                                                         , "CallSurvey.211CSATQ2AgentSatisfaction.count"
                                                         , "Average.CallSurvey.211CSATQ3Recommendation"
                                                         , "CallSurvey.211CSATQ3Recommendation.count"
                                                         , "CallSurvey.CallSurvey.opt_in.count"
                                                         , "Reporting.Queue.Callback.count"),as.double)

CEMonthly_2020 <- CEMonthly_2020 %>% mutate_at(.vars = c("CALLS", "AGENT.count", "Reporting.Abandon_Disposition.count"
                                                         , "ABANDONED.count", "ABANDONED...rec."
                                                         , "SERVICE.LEVEL...rec.", "Average.CallSurvey.211CSATQ1MoreInfo"
                                                         , "CallSurvey.211CSATQ1MoreInfo.count"
                                                         , "Average.CallSurvey.211CSATQ2AgentSatisfaction"
                                                         , "CallSurvey.211CSATQ2AgentSatisfaction.count"
                                                         , "Average.CallSurvey.211CSATQ3Recommendation"
                                                         , "CallSurvey.211CSATQ3Recommendation.count"
                                                         , "CallSurvey.CallSurvey.opt_in.count", "Reporting.Queue.Callback.count"
                                                         ),as.double)

CEMonthly_2021 <- CEMonthly_2021 %>% mutate_at(.vars = c("CALLS", "AGENT.count"
                                                         , "Reporting.Abandon_Disposition.count", "ABANDONED.count", "ABANDONED...rec."
                                                         , "SERVICE.LEVEL...rec."
                                                         , "Average.CallSurvey.211CSATQ1MoreInfo", "CallSurvey.211CSATQ1MoreInfo.count"
                                                         , "Average.CallSurvey.211CSATQ2AgentSatisfaction"
                                                         , "CallSurvey.211CSATQ2AgentSatisfaction.count"
                                                         , "Average.CallSurvey.211CSATQ3Recommendation"
                                                         , "CallSurvey.211CSATQ3Recommendation.count"
                                                         , "CallSurvey.CallSurvey.opt_in.count", "Reporting.Queue.Callback.count"
                                                         ),as.double)
```
Re clean to remove Null COlumns

```{r}
CEMonthly_2019 <- CEMonthly_2019[, colSums(CEMonthly_2019  != 0,na.rm = T) > 0]
CEMonthly_2020 <- CEMonthly_2020[, colSums(CEMonthly_2020  != 0,na.rm = T) > 0]
CEMonthly_2021 <- CEMonthly_2021[, colSums(CEMonthly_2021  != 0,na.rm = T) > 0]
```

Parse Call Type and Skill Type

```{r}
CEMonthly_2019 <- CEMonthly_2019 %>% mutate(CALL_TYPE = str_squish(substr(...2
                                                              ,str_locate(CEMonthly_2019$...2,':')+1
                                                              ,str_locate(CEMonthly_2019$...2,"\\(")-1))
                                            ,LANGUAGE = if_else(str_detect(toupper(...3),"ENGLISH"),"English","Spanish"))

CEMonthly_2020 <- CEMonthly_2020 %>% mutate(CALL_TYPE = str_squish(substr(...2
                                                               ,str_locate(CEMonthly_2020$...2,':')+1
                                                               ,str_locate(CEMonthly_2020$...2,"\\(")-1))
                                            ,LANGUAGE = if_else(str_detect(toupper(...3),"ENGLISH"),"English","Spanish"))

CEMonthly_2021 <- CEMonthly_2021 %>% mutate(CALL_TYPE = str_squish(substr(...2
                                                               ,str_locate(CEMonthly_2021$...2,':')+1
                                                               ,str_locate(CEMonthly_2021$...2,"\\(")-1))
                                            ,LANGUAGE = if_else(str_detect(toupper(...3),"ENGLISH"),"English","Spanish"))
```

Drop intermediate columns for Call Type and Call Skill

```{r}
CEMonthly_2019 <- within(CEMonthly_2019,rm(...2,...3))
CEMonthly_2020 <- within(CEMonthly_2020,rm(...2,...3))
CEMonthly_2021 <- within(CEMonthly_2021,rm(...2,...3))
```


Rename Columns
```{r}
CEMonthly_2019 <- CEMonthly_2019 %>% rename(DATE = DATE
,CALLS = CALLS
,AGENT.COUNT = AGENT.count
,REPORTING.ABANDON_DISPOSITION.COUNT = Reporting.Abandon_Disposition.count
,ABANDONED.COUNT = ABANDONED.count
,ABANDONED.REC = ABANDONED...rec.
,AVERAGE.HANDLE.TIME = Average.HANDLE.TIME
,AVERAGE.SPEED.OF.ANSWER = Average.SPEED.OF.ANSWER
,AVERAGE.QUEUE.WAIT.TIME = Average.QUEUE.WAIT.TIME
,AVERAGE.HOLD.TIME = Average.HOLD.TIME
,SERVICE.LEVEL.REC = SERVICE.LEVEL...rec.
,FILE_NAME = file_name
,SHEET_NAME = sheet_name
,CALL_TYPE = CALL_TYPE
,LANGUAGE = LANGUAGE
)

CEMonthly_2020 <- CEMonthly_2020 %>% rename(DATE = DATE
,CALLS = CALLS
,AGENT.COUNT = AGENT.count
,REPORTING.ABANDON_DISPOSITION.COUNT = Reporting.Abandon_Disposition.count
,ABANDONED.COUNT = ABANDONED.count
,ABANDONED.REC = ABANDONED...rec.
,AVERAGE.HANDLE.TIME = Average.HANDLE.TIME
,AVERAGE.SPEED.OF.ANSWER = Average.SPEED.OF.ANSWER
,AVERAGE.QUEUE.WAIT.TIME = Average.QUEUE.WAIT.TIME
,AVERAGE.HOLD.TIME = Average.HOLD.TIME
,SERVICE.LEVEL.REC = SERVICE.LEVEL...rec.
,FILE_NAME = file_name
,SHEET_NAME = sheet_name
,CALL_TYPE = CALL_TYPE
,LANGUAGE = LANGUAGE

)


CEMonthly_2021 <- CEMonthly_2021 %>% rename(DATE = DATE
,CALLS = CALLS
,AGENT.COUNT = AGENT.count
,REPORTING.ABANDON_DISPOSITION.COUNT = Reporting.Abandon_Disposition.count
,ABANDONED.COUNT = ABANDONED.count
,ABANDONED.REC = ABANDONED...rec.
,AVERAGE.HANDLE.TIME = Average.HANDLE.TIME
,AVERAGE.SPEED.OF.ANSWER = Average.SPEED.OF.ANSWER
,AVERAGE.QUEUE.WAIT.TIME = Average.QUEUE.WAIT.TIME
,AVERAGE.HOLD.TIME = Average.HOLD.TIME
,SERVICE.LEVEL.REC = SERVICE.LEVEL...rec.
,FILE_NAME = file_name
,SHEET_NAME = sheet_name
,CALL_TYPE = CALL_TYPE
,LANGUAGE = LANGUAGE)
```

Write to Output

```{r}
write.csv(CEMonthly_2019,"../Data/Cleaned/CEMonthly_2019.csv",row.names = F)
write.csv(CEMonthly_2020,"../Data/Cleaned/CEMonthly_2020.csv",row.names = F)
write.csv(CEMonthly_2021,"../Data/Cleaned/CEMonthly_2021.csv",row.names = F)
```
