==> Analysis/02_CE211_Analysis.Rmd <==
---
title: "02_CE211_Analysis"
author: "Pratik Chaudhari"
date: "10/27/2021"
output:
  word_document: default
  pdf_document: default
  html_document:
    code_folding: hide
always_allow_html: true
---

Load Libraries
```{r}
require(tidyverse)
require(lubridate)
require(stringr)
require(readxl)
require(arsenal)
require(gtsummary)
require(plotly)
```


```{r}
needs <- read.csv("../Data/Cleaned/01_Needs.csv")
names(needs) <- toupper(names(needs))
referrals <- read.csv("../Data/Cleaned/01_referrals.csv")
names(referrals) <- toupper(names(referrals))
```


```{r message=FALSE, warning=FALSE, echo=FALSE, include = FALSE}
needs <- needs %>% mutate(AGE = as.factor(AGE), Gender = as.factor(GENDER), LANGUAGE = as.factor(LANGUAGE)
                          ,MILITARY_STATUS = as.factor(MILITARY_STATUS)
                          ,DISABILITY_STATUS = as.factor(DISABILITY_STATUS)
                          ,DISABILITY_TYPE = as.factor(DISABILITY_TYPE)
                          ,HEALTH_INSURANCE = as.factor(HEALTH_INSURANCE)
                          );

referrals <- referrals %>% mutate(AGE = as.factor(AGE), Gender = as.factor(GENDER), Language = as.factor(LANGUAGE)
                          ,MILITARY_STATUS = as.factor(MILITARY_STATUS)
                          ,DISABILITY_STATUS = as.factor(DISABILITY_STATUS)
                          ,DISABILITY_TYPE = as.factor(DISABILITY_TYPE)
                          ,HEALTH_INSURANCE = as.factor(HEALTH_INSURANCE))
```





AS per discussions, Housing is the primary need for all clients, so needs variable becomes unnecessary to include.
Removed Needs from the structure and keep unique values
```{r}
needs <- needs %>% select(-c(NEED)) %>% distinct()
```

```{r message=FALSE, warning=FALSE, echo=FALSE}
needs %>% select(CONTACT_NUMBER,CE_SCREENED) %>% distinct() %>% select(CE_SCREENED) %>% tbl_summary()
```


Q1 ----


e. Cross tab: CE Screened (all categories) filtered by Need=Housing  x Gender
Excluding DV - Referral, as it's determined along with some other CE_Screened value
```{r message=FALSE, warning=FALSE, echo=FALSE}
q1.e <- needs %>%  select(CONTACT_NUMBER, CE_SCREENED, GENDER) %>% filter(CE_SCREENED != 'DV Referral') %>% distinct() %>% select(CE_SCREENED, GENDER)
q1.e %>% tbl_summary(by = CE_SCREENED,missing = "no")

```

Calculating for DV referral across calls
```{r message=FALSE, warning=FALSE, echo=FALSE}
q1.e.dv <- needs %>%  select(CONTACT_NUMBER, CE_SCREENED, GENDER) %>% filter(CE_SCREENED == 'DV Referral') %>% distinct()
q1.e.dv <- rbind(q1.e.dv, needs %>%  select(CONTACT_NUMBER, CE_SCREENED, GENDER) %>% filter(!(CONTACT_NUMBER %in% q1.e.dv$CONTACT_NUMBER)) %>% distinct())
q1.e.dv <- q1.e.dv %>% mutate(CE_SCREENED = if_else(CE_SCREENED != 'DV Referral',"No DV Referral",CE_SCREENED)) %>% distinct

q1.e.dv %>% select(CE_SCREENED,GENDER) %>% tbl_summary(by = CE_SCREENED)
q1.e.dv %>% select(CE_SCREENED,GENDER) %>% tbl_summary(by = GENDER)
```

g. Crosstab: CE Screened (all categories) filtered by Need=Housing  x Disability Status

```{r message=FALSE, warning=FALSE, echo=FALSE}
q1.g <- needs %>% select(CONTACT_NUMBER, DISABILITY_STATUS, CE_SCREENED) %>% filter(CE_SCREENED != 'DV Referral') %>% distinct()
q1.g %>% select(DISABILITY_STATUS,CE_SCREENED) %>% tbl_summary(by = CE_SCREENED,missing = "no")

```

Calculating for DV referal across calls
```{r message=FALSE, warning=FALSE, echo=FALSE}
q1.g.dv <- needs %>% select(CONTACT_NUMBER, DISABILITY_STATUS, CE_SCREENED) %>% filter(CE_SCREENED != 'DV Referral') %>% distinct()
q1.g.dv <- rbind(q1.g.dv, needs %>%  select(CONTACT_NUMBER, CE_SCREENED, DISABILITY_STATUS) %>% 
                   filter(!(CONTACT_NUMBER %in% q1.g.dv$CONTACT_NUMBER)) %>% distinct())
q1.g.dv <- q1.g.dv %>% mutate(CE_SCREENED = if_else(CE_SCREENED != 'DV Referral',"No DV Referral",CE_SCREENED)) %>% distinct

q1.g.dv %>% select(CE_SCREENED,DISABILITY_STATUS) %>% tbl_summary(by = CE_SCREENED)
q1.g.dv %>% select(CE_SCREENED,DISABILITY_STATUS) %>% tbl_summary(by = DISABILITY_STATUS)
```


i. Crosstab: CE Screened (all categories) filtered by Need=Housing  x MILITARY_STATUS

```{r message=FALSE, warning=FALSE, echo=FALSE}
q1.i <- needs %>% select(CONTACT_NUMBER, MILITARY_STATUS, CE_SCREENED) %>% filter(CE_SCREENED != 'DV Referral') %>% distinct()
q1.i %>% select(MILITARY_STATUS,CE_SCREENED) %>% tbl_summary(by = CE_SCREENED,missing = "no")

```


```{r message=FALSE, warning=FALSE, echo=FALSE}
CESCreened_REferal <- needs %>% select(CONTACT_NUMBER,CE_SCREENED,REFERRAL) %>% distinct()
CESCreened_REferal %>% select(CE_SCREENED,REFERRAL) %>% tbl_summary(by=CE_SCREENED,missing = 'no',sort = list(everything() ~ "frequency"))
```



Q2 - 

```{r message=FALSE, warning=FALSE, echo=FALSE}
q2 <- needs %>% select(CONTACT_NUMBER,GENDER,DISABILITY_STATUS,MILITARY_STATUS,AGE,REFERRAL) %>% distinct()
```

Q2.1 -- REFERAL - GENDER
```{r}
q2 %>% select(GENDER,REFERRAL) %>% tbl_summary(by = GENDER,missing = "no")
```

Q2.2 -- REFERAL - DISABILITY_STATUS
```{r}
q2 %>% select(DISABILITY_STATUS,REFERRAL) %>% tbl_summary(by = DISABILITY_STATUS,missing = "no")
```

Q2.3 -- REFERAL - MILITARY_STATUS
```{r}
q2 %>% select(MILITARY_STATUS,REFERRAL) %>% tbl_summary(by = MILITARY_STATUS,missing = "no")
```
